feat(password-recovery): implement password recovery logic
- Add custom hook for password recovery logic
- Refactor PasswordRecoveryScreen to use the new hook
- Update imports and remove unused files
- Ensure password validation and error handling in the hook
