--- src/features/password-recovery/index.ts
@@ -1,2 +1,0 @@
-export * from './utils';
-export * from './types';
