+++ src/features/password-recovery/types.ts
@@ -1,4 +1,4 @@
 import * as z from 'zod/mini';
-import { PasswordRecoverySchema } from './utils';
+import { PasswordRecoverySchema } from './utils/schema';
 
 export type PasswordRecoverySchemaType = z.infer<typeof PasswordRecoverySchema>;
