--- src/features/password-recovery/utils/index.ts
@@ -1,1 +1,0 @@
-export * from './schema';
